--- src/components/Top.tsx
@@ -1,24 +1,0 @@
-function Top() {
-  return (
-    <>
-      <nav className="top">
-        <ul className="top__list">
-          <li>
-            <a href="#">Доставка і оплата</a>
-          </li>
-          <li>
-            <a href="#">Про магазин</a>
-          </li>
-          <li>
-            <a href="#">Відгуки</a>
-          </li>
-          <li>
-            <a href="#">Контакти</a>
-          </li>
-        </ul>
-      </nav>
-    </>
-  )
-}
-
-export default Top
