+++ src/sass/layout/Top.tsx
@@ -1,0 +1,26 @@
+function Top() {
+  return (
+    <>
+      <div className="container">
+        <nav className="top">
+          <ul className="top__list">
+            <li>
+              <a href="#">Доставка і оплата</a>
+            </li>
+            <li>
+              <a href="#">Про магазин</a>
+            </li>
+            <li>
+              <a href="#">Відгуки</a>
+            </li>
+            <li>
+              <a href="#">Контакти</a>
+            </li>
+          </ul>
+        </nav>
+      </div>
+    </>
+  )
+}
+
+export default Top
